\name{calcTGIF}
\alias{calcTGIF}
\docType{methods}
\alias{calcTGIF,SingleCellExperiment-method}
\alias{calcTGIF}
\title{
Function for connecting cellular patterns and functional patterns using joint NMF
}
\description{
\code{\link{calcTGIF}} function calculates what kind of cellular patterns and functional patterns are contained in single-cell RNA-seq data and \code{\link{reportTGIF}} function generates report of analytic result.
}
\usage{
calcTGIF(sce, rank)
}
\arguments{
\item{sce}{
A object generated by instantization of SingleCellExperiment-class.
}
\item{rank}{
Rank parameter of joint NMF algorithm.
}
}
\value{
The result is saved to metadata slot of SingleCellExperiment object.
}
\author{
\packageAuthor{scTGIF}
}
\examples{
showMethods("calcTGIF")
}

\keyword{methods}