\name{label.DistalLungEpithelium}
\alias{label.DistalLungEpithelium}
\docType{data}
\title{
Cellular label of DistalLungEpithelium dataset containing five cluster.
}
\description{
A vector containing 80 elements (cells).
}
\usage{data("label.DistalLungEpithelium")}

\references{
Treutlein, B. et al. (2014) Reconstructing lineage hierarchies of the distal lung epithelium using single-cell RNA-seq. \emph{Nature} \bold{509}, 371-375
}
\examples{
data("label.DistalLungEpithelium")
}
\keyword{datasets}