\name{pca.DistalLungEpithelium}
\alias{pca.DistalLungEpithelium}
\docType{data}
\title{
The result of PCA of the DistalLungEpithelium dataset.
}
\description{
A matrix having 80 (cells) * 2 (PCs) elements.
}
\usage{data("pca.DistalLungEpithelium")}

\references{
Treutlein, B. et al. (2014) Reconstructing lineage hierarchies of the distal lung epithelium using single-cell RNA-seq. \emph{Nature} \bold{509}, 371-375
}
\examples{
data("pca.DistalLungEpithelium")
}
\keyword{datasets}