\name{scTGIF-package}
\alias{scTGIF-package}
\alias{scTGIF}
\docType{package}
\title{
\packageTitle{scTGIF}
}
\description{
\packageDescription{scTGIF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scTGIF}
\packageIndices{scTGIF}
\code{\link{calcTGIF}} function calculates what kind of cellular patterns and functional patterns are contained in single-cell RNA-seq data and \code{\link{reportTGIF}} function generates report of analytic result.
}
\author{
\packageAuthor{scTGIF}

Maintainer: \packageMaintainer{scTGIF}
}
\references{
Diego Adhemar Jaitin, Ephraim Kenigsberg, Hadas Keren-Shaul, Naama Elefant, Franziska Paul, Irina Zaretsky, Alexander Mildner, Nadav Cohen, Steffen Jung, Amos Tanay, Ido Amit (2014) Massively Parallel Single-Cell RNA-Seq for Marker-Free Decomposition of Tissues into Cell Types. \emph{Science}, \bold{343 (6172)}: 776-779

Florian Buettner, Kedar N Natarajan, F Paolo Casale, Valentina Proserpio, Antonio Scialdone, Fabian J Theis, Sarah A Teichmann, John C Marioni, Oliver Stegle (2015) Computational analysis of cell-to-cell heterogeneity in single-cell RNA-sequencing data reveals hidden subpopulations of cells. \emph{Nature Biotechnology}, \bold{33}, 155-160

Dominic Grun, Anna Lyubimova, Lennart Kester, Kay Wiebrands,  Onur Basak, Nobuo Sasaki, Hans Clevers, Alexander van Oudenaarden (2015) Single-cell messenger RNA sequencing reveals rare intestinal cell types. \emph{Nature}, \bold{525}: 251-255
}
\keyword{ package }
