\name{calcTGIF}
\alias{calcTGIF}
\docType{methods}
\title{
Function for calculating cellular patterns and functional patterns
}
\description{
\code{\link{calcTGIF}} function calculates what kind of cellular patterns and functional patterns are contained in single-cell RNA-seq data and \code{\link{reportTGIF}} function generates report of analytic result.
}
\usage{
calcTGIF(DataMatrix = NA, gmt = NA)
}
\arguments{
  \item{DataMatrix}{
Data matrix user want to analyze. Confirm that your DataMatrix has Human Entrez Gene ID.
}
  \item{gmt}{
Object generated from GSEABase::getGmt function.

GMT file can be downloaded from MSigDB web (site http://software.broadinstitute.org/gsea/login.jsp#msigdb).

Please confirm that the gmt file contains Human Entrez Gene ID, not gene symbol. Also confirm that the DataMatrix has Human Entrez Gene ID.

}
}
\value{
\item{d}{D Matrix generated by SVD}
\item{u}{U matrix generated by SVD}
\item{v}{V matrix generated by SVD}
\item{NoCell}{Number of Cell}
\item{NoGene}{Number of Gene}
\item{NoTerm}{Number of Functional Term}
\item{NameCell}{Name of Cell}
\item{NameGene}{Name of Human Entrez Gene ID}
\item{NameTerm}{Name of Functional Term}
}
\author{
\packageAuthor{scTGIF}
}

\examples{
\dontrun{
    # Load the GSEABase for using getGmt function
    library("scTGIF")
    library("GSEABase")

    # Load the gmt file from MSigDB
    # Only "Entrez Gene ID" can be used in scTGIF
    gmt <- GSEABase::getGmt("/PATH/YOU/SAVED/THE/GMT/h.all.v6.0.entrez.gmt")

    # Test data
    load("MARS.RData")

    # Perform calcTGIF
    tgif.MARS <- calcTGIF(exprs(eSet.MARS), gmt)

    # Report the result of calcTGIF
    reportTGIF(result=tgif.MARS, Comp.Matrix=pData(eSet.MARS)$tSNE.2D, nDim=10, label=pData(eSet.MARS)$Color, out.dir=".", html.open=TRUE, title="scTGIF Report for MARS-Seq Data (t-SNE, 2D)", author="Koki Tsuyuzaki")
    }
}