\name{calcTGIF}
\alias{calcTGIF}
\docType{methods}
\alias{calcTGIF,SingleCellExperiment-method}
\alias{calcTGIF}
\title{
Function for connecting cellular patterns and functional patterns using jNMF
}
\description{
\code{\link{calcTGIF}} function calculates what kind of
cellular patterns and functional patterns are contained in
single-cell RNA-seq data and \code{\link{reportTGIF}} function
generates report of analytic result.
}
\usage{
calcTGIF(sce, ndim, verbose=FALSE, droplet=TRUE)
}
\arguments{
\item{sce}{
A object generated by instantization of SingleCellExperiment-class.
}
\item{ndim}{
The number of low-dimension of joint NMF algorithm.
}
\item{verbose}{
The verbose parameter for nnTensor::jNMF (Default: FALSE).
}
\item{droplet}{
Whether Droplet-based single-cell RNA-Seq or not (Default: TRUE).
}
}
\value{
The result is saved to metadata slot of SingleCellExperiment object.
}
\author{
\packageAuthor{scTGIF}
}
\examples{
showMethods("calcTGIF")
}

\keyword{methods}