\name{scTGIF-package}
\alias{scTGIF-package}
\alias{scTGIF}
\docType{package}
\title{
\packageTitle{scTGIF}
}
\description{
\packageDescription{scTGIF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scTGIF}
\packageIndices{scTGIF}
\code{\link{calcTGIF}} function calculates what kind of cellular patterns and functional patterns are contained in single-cell RNA-seq data and \code{\link{reportTGIF}} function generates report of analytic result.
The algorithm is based on joint NMF, which is implemented in nnTensor package.
}
\author{
\packageAuthor{scTGIF}

Maintainer: \packageMaintainer{scTGIF}
}
\references{
Dominic Grun, Anna Lyubimova, Lennart Kester, Kay Wiebrands,  Onur Basak, Nobuo Sasaki, Hans Clevers, Alexander van Oudenaarden (2015) Single-cell messenger RNA sequencing reveals rare intestinal cell types. \emph{Nature}, \bold{525}: 251-255
}
\keyword{ package }
