\name{settingTGIF}
\alias{settingTGIF}
\docType{methods}
\alias{settingTGIF,SingleCellExperiment-method}
\alias{settingTGIF}
\title{
Paramter setting for scTGIF
}
\description{
All parameters is saved to metadata slot of SingleCellExperiment object.
}
\usage{
settingTGIF(sce, gmt, reducedDimNames, assayNames="counts", nbins=40)
}
\arguments{
\item{sce}{
A object generated by instantization of SingleCellExperiment-class.
}
\item{gmt}{
Object generated from GSEABase::getGmt function.
GMT file can be downloaded from MSigDB web
(site http://software.broadinstitute.org/gsea/login.jsp#msigdb).
Please confirm that the gmt file contains Human Entrez Gene ID,
not gene symbol. Also confirm that the DataMatrix has Human Entrez Gene ID.
}
\item{reducedDimNames}{
The names of reducedDim(sce) that user want use in scTGIF.
}
\item{assayNames}{
The unit of gene expression for using scTGIF
(e.g. normcounts, cpm...etc) (Default: "counts").
}
\item{nbins}{
The number of bins of schex (Default: 40).
}
}
\value{
The result is saved to metadata slot of SingleCellExperiment object.
}
\author{
\packageAuthor{scTGIF}
}
\examples{
showMethods("settingTGIF")
}

\keyword{methods}